changes functionalities
